medai_compass/models/path_foundation.py
"""Path Foundation model wrapper for histopathology.

Provides:
- 384-dimensional embedding extraction from WSI patches
- Batch processing support
"""

from typing import Optional
import numpy as np

# Lazy imports
torch = None


def _lazy_import_torch():
    """Lazily import torch."""
    global torch
    if torch is None:
        import torch as _torch
        torch = _torch


class PathFoundationWrapper:
    """
    Wrapper for Google's Path Foundation model.
    
    Extracts 384-dimensional embeddings from histopathology patches
    for downstream classification and analysis tasks.
    """
    
    def __init__(
        self,
        model_name: str = "google/path-foundation",
        device: Optional[str] = None
    ):
        """
        Initialize Path Foundation wrapper.
        
        Args:
            model_name: Model identifier
            device: Device to use (cuda/cpu/auto)
        """
        _lazy_import_torch()
        
        self.model_name = model_name
        
        # Determine device
        if device is None:
            device = "cuda" if torch.cuda.is_available() else "cpu"
        self.device = device
        
        # Load model from torch hub (placeholder for actual loading)
        self.model = torch.hub.load(
            "google/health-ai-developer-foundations",
            "path_foundation",
            pretrained=True
        )
        self.model = self.model.to(self.device)
        self.model.eval()
        
        # Standard preprocessing for Path Foundation
        self.patch_size = 224
        self.embedding_dim = 384
    
    def preprocess(self, image: np.ndarray) -> "torch.Tensor":
        """
        Preprocess image patch for Path Foundation.
        
        Args:
            image: RGB image array (H, W, 3)
            
        Returns:
            Preprocessed tensor
        """
        from PIL import Image
        from torchvision import transforms
        
        # Standard ImageNet normalization
        transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(
                mean=[0.485, 0.456, 0.406],
                std=[0.229, 0.224, 0.225]
            )
        ])
        
        pil_image = Image.fromarray(image)
        tensor = transform(pil_image).unsqueeze(0)
        
        return tensor.to(self.device)
    
    def get_embedding(self, image: np.ndarray) -> np.ndarray:
        """
        Extract embedding from a single image patch.
        
        Args:
            image: RGB image array (H, W, 3)
            
        Returns:
            384-dimensional embedding vector
        """
        tensor = self.preprocess(image)
        
        with torch.no_grad():
            embedding = self.model(tensor)
        
        return embedding.detach().cpu().numpy().squeeze()
    
    def get_embeddings_batch(
        self, 
        images: list[np.ndarray],
        batch_size: int = 32
    ) -> np.ndarray:
        """
        Extract embeddings from multiple image patches.
        
        Args:
            images: List of RGB image arrays
            batch_size: Processing batch size
            
        Returns:
            Array of embeddings (N, 384)
        """
        all_embeddings = []
        
        for i in range(0, len(images), batch_size):
            batch_images = images[i:i + batch_size]
            
            # Preprocess batch
            tensors = [self.preprocess(img) for img in batch_images]
            batch_tensor = torch.cat(tensors, dim=0)
            
            with torch.no_grad():
                embeddings = self.model(batch_tensor)
            
            all_embeddings.append(embeddings.detach().cpu().numpy())
        
        return np.concatenate(all_embeddings, axis=0)
